requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-svg==0.0.12
dash-table==5.0.0
Flask==2.2.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pip==23.0.1
plotly==5.14.1
python-dotenv==1.0.0
requests==2.28.2
setuptools==66.0.0
tenacity==8.2.2
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
